# candidate-match/openapi.yaml
openapi: 3.0.3
info:
  title: Cloudberries Candidate Match API
  version: 1.1.0
  description: |
    API for candidate matching, CV retrieval, embeddings, and health checks.
    Includes normalized health responses and skill-based matching endpoint.
    Generated from controller sources as of this commit.
  contact:
    name: Cloudberries
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/project-requests:
    get:
      tags: [ Project Requests ]
      summary: List stored customer project requests with pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          description: Page number (0-based)
        - in: query
          name: size
          schema:
            type: integer
            default: 20
          description: Number of items per page
        - in: query
          name: sort
          schema:
            type: string
            default: "id,desc"
          description: Sort specification (field,direction)
      responses:
        '200':
          description: Paged project requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedProjectRequestResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [ Project Requests ]
      summary: Create a new project request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequestDto'
            example:
              customerName: "Acme Corp AS"
              requiredSkills: ["KOTLIN", "JAVA"]
              startDate: "2024-04-01T09:00:00"
              endDate: "2024-10-01T17:00:00"
              responseDeadline: "2024-03-15T17:00:00"
              requestDescription: "We need a senior backend developer for a 6-month project using Kotlin and Spring Boot."
              responsibleSalespersonEmail: "sales@acme.com"
      responses:
        '201':
          description: Project request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestDto'
        '400':
          description: Invalid request data
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/project-requests/upload:
    post:
      tags: [ Project Requests ]
      summary: Upload and analyze a customer project request PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [ file ]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Extracted project request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/project-requests/{id}:
    get:
      tags: [ Project Requests ]
      summary: Get stored customer project request by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestResponseDto'
        '404':
          description: Not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/project-requests/{id}/close:
    put:
      tags: [ Project Requests ]
      summary: Close a project request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project request closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestDto'
        '404':
          description: Project request not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/project-requests/{id}/analyze:
    post:
      tags: [ Project Requests ]
      summary: Trigger AI analysis for a project request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestDto'
        '404':
          description: Project request not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/project-requests/{id}/suggestions:
    get:
      tags: [ Project Requests ]
      summary: Get AI suggestions for a project request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: AI suggestions for the project request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AISuggestionDto'
        '404':
          description: Project request not found
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    ProjectRequestResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerName:
          type: string
        title:
          type: string
        summary:
          type: string
        originalFilename:
          type: string
        mustRequirements:
          type: array
          items:
            $ref: '#/components/schemas/ProjectRequirementDto'
        shouldRequirements:
          type: array
          items:
            $ref: '#/components/schemas/ProjectRequirementDto'
    ProjectRequirementDto:
      type: object
      required: [ name ]
      properties:
        name:
          type: string
        details:
          type: string
    PagedProjectRequestResponseDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProjectRequestResponseDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer
        pageSize:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    CreateProjectRequestDto:
      type: object
      required: [ customerName, requiredSkills, startDate, endDate, responseDeadline, requestDescription, responsibleSalespersonEmail ]
      properties:
        customerName:
          type: string
          minLength: 1
        requiredSkills:
          type: array
          items:
            type: string
            enum: [ KOTLIN, JAVA, PYTHON, JAVASCRIPT, TYPESCRIPT, REACT, ANGULAR, VUE, SPRING_BOOT, BACKEND, FRONTEND, FULLSTACK, AZURE, AWS, DOCKER, KUBERNETES ]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        responseDeadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [ OPEN, IN_PROGRESS, CLOSED ]
          default: OPEN
        requestDescription:
          type: string
          minLength: 1
        responsibleSalespersonEmail:
          type: string
          format: email
    ProjectRequestDto:
      type: object
      required: [ customerName, requiredSkills, startDate, endDate, responseDeadline, status, requestDescription, responsibleSalespersonEmail ]
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        customerName:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        responseDeadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [ OPEN, IN_PROGRESS, CLOSED ]
        requestDescription:
          type: string
        responsibleSalespersonEmail:
          type: string
          format: email
        aiSuggestions:
          type: array
          items:
            $ref: '#/components/schemas/AISuggestionDto'
    AISuggestionDto:
      type: object
      required: [ consultantName, userId, cvId, matchScore, justification, createdAt ]
      properties:
        id:
          type: integer
          format: int64
        consultantName:
          type: string
        userId:
          type: string
        cvId:
          type: string
        matchScore:
          type: number
          format: double
          minimum: 0.0
          maximum: 100.0
        justification:
          type: string
        createdAt:
          type: string
          format: date-time
        skills:
          type: array
          items:
            type: string
...
    Problem:
      type: object
      description: Generic error payload
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string